fix(ci): resolve -Werror warnings under clang/gcc and Qt 6.11
- Add a constructor to EventWrapper (aggregate init was missing fields)
- Fix sign-conversion/shadow/warning-flag violations in dom_binding, layout,
  browser, renderer, ui and style
- ElementGeometry now uses float (matching the layout tree) with explicit
  double conversions at the JS boundary
- UI tests use QTest instead of the deprecated QMouseEvent constructors;
  link Qt6::Test
- Suppress a GCC false-positive array-bounds warning in the SVG rasterizer

+++ src/renderer/include/neko/renderer/page.h
@@ -29,14 +29,14 @@ namespace neko::renderer {
 // |border_top|/|border_left| the top/left border widths.
 struct ElementGeometry
 {
-  double x = 0;
-  double y = 0;
-  double width = 0;
-  double height = 0;
-  double client_width = 0;
-  double client_height = 0;
-  double border_top = 0;
-  double border_left = 0;
+  float x = 0;
+  float y = 0;
+  float width = 0;
+  float height = 0;
+  float client_width = 0;
+  float client_height = 0;
+  float border_top = 0;
+  float border_left = 0;
 };
 
 // The minimal page pipeline: HTML -> DOM -> style -> layout -> paint.
